requirements.txt
boto3
boto
Flask==1.02
Flask-Cors==3.0.2
gunicorn==19.0.0
Jinja2==2.9.4
numexpr==2.6.2
numpy==1.15.4
oauthlib==1.0.3
Pillow==5.3.0
requests==2.21.0
requests-file==1.4.1
requests-oauthlib==0.6.1
scikit-image
scipy
torch=1.0.0
torchvision
matplotlib
opencv-python